% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_probability.R
\name{capture_probability_estimate}
\alias{capture_probability_estimate}
\title{Get the capture probability estimate given the specified estimator type.}
\usage{
capture_probability_estimate(qhat_list, estimator = c("PI", "DR", "TMLE"))
}
\description{
Get the capture probability estimate given the specified estimator type.
}
