% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare Simulation Results Against Ground Truth}
\usage{
compare(sim, ground_truth)
}
\arguments{
\item{sim}{list: simulations produced by \code{simulate}}

\item{ground_truth}{list: ground truth values produced by
\code{extract_ground_truth}}
}
\value{
tibble with AIC, estimate, confidence intervals, and ground truth
for each group
}
\description{
Compare Simulation Results Against Ground Truth
}
