% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crc_aic.R
\name{crc_aic}
\alias{crc_aic}
\title{Run CRC with AIC Formula Selection. This function takes in the data and a
list of potential formulas to fit, fits the specified model for each
formula, and returns a dataframe with the formula, estimate, AIC value,
confidence interval, and any error messages for each formula.}
\usage{
crc_aic(data, opts)
}
\arguments{
\item{data}{a data frame containing the observed capture histories and a
frequency column.}

\item{formulas}{a list of formula objects specifying the log-linear models to
fit and compare using AIC}

\item{model_family}{a string specifying the model family to use, either
"poisson" or "negbin". Default is "poisson".}
}
\value{
a data frame with the formula, estimate, AIC value, confidence
interval, and any error messages for each formula
}
\description{
Run CRC with AIC Formula Selection. This function takes in the data and a
list of potential formulas to fit, fits the specified model for each
formula, and returns a dataframe with the formula, estimate, AIC value,
confidence interval, and any error messages for each formula.
}
\keyword{internal}
