% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crc_plugin.R
\name{crc_plugin}
\alias{crc_plugin}
\title{CRC with Plugin Estimators}
\usage{
crc_plugin(data, opts)
}
\arguments{
\item{data}{a data frame containing the capture data. All columns must be
numeric type as long as we use the \code{drpop} package.}

\item{opts}{a \code{PluginOptions} object specifying the plugin method and
parameters for the TMLE estimation.}
}
\value{
a list containing the TMLE estimate of population size and
confidence intervals.
}
\description{
This function runs CRC using a plugin estimator for the
nuisance functions.
The plugin estimator is specified in the \code{PluginOptions} object passed
to the \code{opts} argument. Supported plugin methods are "logit", "ranger",
"logitranger", and "gam". The function will call the appropriate estimation
function based on the specified plugin method and return the TMLE estimate
of population size and confidence intervals.
}
