% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crc_stepwise.R
\name{crc_stepwise}
\alias{crc_stepwise}
\title{Run CRC with Stepwise Formula Selection.}
\usage{
crc_stepwise(data, opts)
}
\arguments{
\item{data}{a data frame containing the observed capture histories and a
frequency column.}

\item{opts}{a \code{StepwiseOptions} object specifying the options for
stepwise regression, including the model family, p-value threshold for
variable inclusion, and stepwise direction.}
}
\description{
This function runs CRC using stepwise regression for formula
selection. It takes in the data and options for stepwise regression, fits the
specified model using stepwise regression, and returns a list with the final
formula, estimate, confidence interval, and AIC value for the selected model.
}
\keyword{internal}
