% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossfit_fold.R
\name{crossfit_fold}
\alias{crossfit_fold}
\title{Create cross-validation fold assignments for CRC estimation}
\usage{
crossfit_fold(data, nfolds)
}
\arguments{
\item{data}{A data frame containing the capture data. This should be the
raw' capture data, not a frequency table.}

\item{nfolds}{Integer scalar giving the number of cross-validation folds to
create. Must be a positive integer.}
}
\value{
An integer vector of the same length as the number of rows in
\code{data}, where each element is an integer from 1 to \code{nfolds}
indicating the fold assignment for that row.
}
\description{
Create cross-validation fold assignments for CRC estimation
}
