% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_fit.R
\name{empirical_qhats}
\alias{empirical_qhats}
\title{Empirical mean for each capture list and their product for the provided j
and k indices. This is used as a fallback when there are no covariates (i.e.
when the number of columns in train is less than or equal to n_capture_cols).}
\usage{
empirical_qhats(train, test, j, k)
}
\arguments{
\item{train}{Training data frame}

\item{test}{Test data frame}

\item{j}{Index of first capture list}

\item{k}{Index of second capture list}
}
\value{
List with elements q1, q2, q12 (predicted capture probabilities)
}
\description{
Empirical mean for each capture list and their product for the provided j
and k indices. This is used as a fallback when there are no covariates (i.e.
when the number of columns in train is less than or equal to n_capture_cols).
}
\keyword{internal}
