% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control.R
\name{extract_captured_data}
\alias{extract_captured_data}
\title{Get Captured Rows}
\usage{
extract_captured_data(data_table, capture)
}
\arguments{
\item{data_table}{data.frame: Data frame containing either a frequency table
or individual records with capture columns.}

\item{capture}{character vector: names of capture columns}
}
\value{
tibble containing only rows where people are captured.
}
\description{
Function to filter out rows where all capture columns are not 0.
This is used to identify all the captured rows in a dataset.
}
