% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_ground_truth}
\alias{extract_ground_truth}
\title{Extract Uncaptured Ground Truth}
\usage{
extract_ground_truth(
  data_table,
  group,
  capture = c("APCD", "BSAS", "Casemix", "Death", "Matris", "PMP")
)
}
\arguments{
\item{data_table}{data.frame: contingency table from \code{create_data}}

\item{group}{character: strata column used to name output values}

\item{capture}{character vector: names of capture columns}
}
\value{
named list of uncaptured frequency counts
}
\description{
Extracts the uncaptured count from a contingency table where uncaptured
observations are defined as rows with all capture columns equal to 0.
}
