% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control.R
\name{extract_uncaptured_data}
\alias{extract_uncaptured_data}
\title{Get Uncaptured Rows}
\usage{
extract_uncaptured_data(data_table, capture)
}
\arguments{
\item{data_table}{data.frame: Data frame containing either a frequency table
or individual records with capture columns.}

\item{capture}{character vector: names of capture columns}
}
\value{
tibble containing only rows where people are uncaptured.
}
\description{
Function to filter out rows where all capture columns are 0.
This is used to identify all the uncaptured rows in a dataset, which is
necessary for extracting a ground truth value for simulation evaluation.
}
