% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{filter_observed_rows}
\alias{filter_observed_rows}
\title{Filter Contingency Table to Observed Rows. This returns only rows where all
capture variables are 0. Thus, the "unknown" rows are removed and the rows
with at least one observation are kept.}
\usage{
filter_observed_rows(data_table, capture)
}
\arguments{
\item{data_table}{data.table: the contingency table to filter}

\item{capture}{list: strings of captures}
}
\description{
Filter Contingency Table to Observed Rows. This returns only rows where all
capture variables are 0. Thus, the "unknown" rows are removed and the rows
with at least one observation are kept.
}
\keyword{internal}
