% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_fit.R
\name{fit_gam}
\alias{fit_gam}
\title{Fit a generalized additive model to the capture list given the covariates.
This is a helper function for gam_qhats to avoid code repetition.}
\usage{
fit_gam(data, formula)
}
\arguments{
\item{data}{Data frame containing the capture columns and covariates.}

\item{i}{Index of the capture column to fit.}

\item{n}{Number of capture columns (the first n columns are capture columns).}
}
\value{
Fitted GAM model.
}
\description{
Fit a generalized additive model to the capture list given the covariates.
This is a helper function for gam_qhats to avoid code repetition.
}
\keyword{internal}
