% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_fit.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit a logistic regression model to the capture list given the covariates.
This is a helper function for logit_qhats to avoid code repetition.}
\usage{
fit_glm(data, formula)
}
\arguments{
\item{data}{Data frame containing the capture columns and covariates.}

\item{i}{Index of the capture column to fit.}

\item{n}{Number of capture columns (the first n columns are capture columns).}
}
\value{
Fitted logistic regression model.
}
\description{
Fit a logistic regression model to the capture list given the covariates.
This is a helper function for logit_qhats to avoid code repetition.
}
\keyword{internal}
