% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_fit.R
\name{fit_random_forest}
\alias{fit_random_forest}
\title{Fit a random forest model to the capture list given the covariates.
This is a helper function for random_forest_qhats to avoid code repetition.}
\usage{
fit_random_forest(data, formula)
}
\arguments{
\item{data}{Data frame containing the capture columns and covariates.}

\item{i}{Index of the capture column to fit.}

\item{n}{Number of capture columns (the first n columns are capture columns).}

\item{...}{Additional arguments passed to ranger().}
}
\value{
Fitted random forest model.
}
\description{
Fit a random forest model to the capture list given the covariates.
This is a helper function for random_forest_qhats to avoid code repetition.
}
\keyword{internal}
