% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.R
\name{gam_estimate}
\alias{gam_estimate}
\title{TMLE using generalized additive model regression
This function performs TMLE estimation of population size using generative
additive models as the nuisance function estimator. It is currently a
wrapper for the drpop package.}
\usage{
gam_estimate(data, nfolds = 2, margin = 0.005)
}
\arguments{
\item{data}{a data frame containing the capture data. All columns must be
numeric type as long as we use the \code{drpop} package.}

\item{nfolds}{int: number of folds to use for cross-fitting. Default is 2.}

\item{margin}{numeric: margin parameter for the TMLE estimation. Default is
0.005.}
}
\value{
a list containing the TMLE estimate of population size and
confidence intervals.
}
\description{
TMLE using generalized additive model regression
This function performs TMLE estimation of population size using generative
additive models as the nuisance function estimator. It is currently a
wrapper for the drpop package.
}
