% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_probability.R
\name{get_double_robust_estimation}
\alias{get_double_robust_estimation}
\title{Doubly-robust estimator for inverse capture}
\usage{
get_double_robust_estimation(q_j, q_k, q_jk, y_j, y_k)
}
\arguments{
\item{qhat_list}{List with elements q_j, q_k, q_jk (numeric vectors)}

\item{pair}{Character or vector identifying the pair
(not used in calculation)}
}
\value{
List with n_hat, sd_n_hat, psi_inverse_hat, se_psi_inverse_hat
}
\description{
Doubly-robust estimator for inverse capture
}
