% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_probability.R
\name{get_plugin_estimation}
\alias{get_plugin_estimation}
\title{Get the capture probability estimate for the standard Plugin Estimator.}
\usage{
get_plugin_estimation(q_j, q_k, q_jk)
}
\arguments{
\item{qhat_list}{List with elements q_j, q_k, q_jk (numeric vectors)}
}
\value{
The capture probability estimate (psi_hat)
}
\description{
Get the capture probability estimate for the standard Plugin Estimator.
}
