% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{is_formula}
\alias{is_formula}
\title{Formula Objects
Generic test of an object being interpretable as a formula}
\usage{
is_formula(x)
}
\arguments{
\item{x}{object to be tested}
}
\description{
Formula Objects
Generic test of an object being interpretable as a formula
}
