% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{is_frequency_table}
\alias{is_frequency_table}
\title{A function used to check if the provided data is a frequency table. This is
necessary as the CRC estimation functions require a frequency table as input.}
\usage{
is_frequency_table(data, frequency_column)
}
\arguments{
\item{data}{a data frame}

\item{frequency_column}{a string specifying the name of the frequency column
in the data frame}
}
\value{
a boolean indicating whether the data frame is a frequency table
(i.e., contains the specified frequency column and that column is numeric)
}
\description{
A function used to check if the provided data is a frequency table. This is
necessary as the CRC estimation functions require a frequency table as input.
}
\keyword{internal}
