% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petersen_estimate.R
\name{petersen_estimate}
\alias{petersen_estimate}
\title{Compute Petersen estimates for all pairs of capture indicators in the data.}
\usage{
petersen_estimate(data)
}
\arguments{
\item{data}{A data frame containing binary capture indicators as columns.}
}
\value{
A list of Petersen estimates and standard errors for each pair of
capture indicators.
}
\description{
This function takes a data frame with binary capture indicators as columns
and computes Petersen estimates and standard errors for each pair of capture
indicators. It returns a list of results for each pair.
}
