% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petersen_estimate.R
\name{petersen_pair_estimate}
\alias{petersen_pair_estimate}
\title{Compute Petersen estimates for all pairs of capture indicators in the data.}
\usage{
petersen_pair_estimate(q1, q2, q12, n_obs)
}
\arguments{
\item{q1}{Vector of predicted capture probabilities for list j.}

\item{q2}{Vector of predicted capture probabilities for list k.}

\item{q12}{Vector of predicted joint capture probabilities for lists j and k.}

\item{n_obs}{Number of observations in the test set.}
}
\value{
A list of Petersen estimates and standard errors for each pair of
capture indicators.
}
\description{
This function takes an estimated q1, q2, q12 for each pair of capture
indicators and computes Petersen estimates and standard errors for each
pair. It returns a list of results for each pair.
}
\keyword{internal}
