% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_fit.R
\name{predict_ranger_model_qhat}
\alias{predict_ranger_model_qhat}
\title{Predict capture probabilities from a fitted ranger model.
This is a helper function for random_forest_qhats to avoid code repetition.
Annoyingly, ranger expects a `data` argument rather than `newdata`, and the
predict function for ranger models returns a list with "predictions" element
that contains the predicted probabilities, so we need a separate function to
extract those probabilities.}
\usage{
predict_ranger_model_qhat(model, data)
}
\arguments{
\item{model}{Fitted ranger model.}

\item{data}{Data frame.}
}
\value{
Predicted probabilities from the specified ranger model.
}
\description{
Predict capture probabilities from a fitted ranger model.
This is a helper function for random_forest_qhats to avoid code repetition.
Annoyingly, ranger expects a `data` argument rather than `newdata`, and the
predict function for ranger models returns a list with "predictions" element
that contains the predicted probabilities, so we need a separate function to
extract those probabilities.
}
\keyword{internal}
