% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_fit.R
\name{predict_stats_model_qhat}
\alias{predict_stats_model_qhat}
\title{Predict capture probabilities from a fitted model.
This is a helper function for the qhat_* functions to avoid code repetition.}
\usage{
predict_stats_model_qhat(model, data)
}
\arguments{
\item{model}{Fitted model.}

\item{data}{Data frame.}
}
\value{
Predicted probabilities from the specified model.
}
\description{
Predict capture probabilities from a fitted model.
This is a helper function for the qhat_* functions to avoid code repetition.
}
\keyword{internal}
