% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Attempts recovery of the ground truth - fundamentally a wrapper for crc}
\usage{
simulate(
  data_table,
  group,
  suppress,
  method,
  formula_selection,
  opts_stepwise,
  capture = c("APCD", "BSAS", "Casemix", "Death", "Matris", "PMP")
)
}
\arguments{
\item{data_table}{data.table from the \code{create.data} step}

\item{group}{stratification string to group by}

\item{suppress}{numeric: maximum value to suppress at}

\item{method}{string: selection for spatial capture recapture model - either
"poisson" or "negbin"}

\item{formula_selection}{string: selection for formula decision - either
"aic", "corr", or "stepwise"}

\item{opts_stepwise}{list: list containing 'direction' of 'forward'
'backward' or 'both', and 'threshold': p value threshold for stepwise
selection}

\item{capture}{list: strings of captures}
}
\description{
Attempts recovery of the ground truth - fundamentally a wrapper for crc
}
\keyword{internal}
