% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_regression.R
\name{step_regression}
\alias{step_regression}
\title{Helper function for stepwise regression}
\usage{
step_regression(
  data,
  y,
  x,
  method = "poisson",
  direction = "both",
  p_threshold = 0.05,
  k = 2,
  verbose = TRUE
)
}
\arguments{
\item{data}{dataframe}

\item{y}{string: LHS of formula object}

\item{x}{string: RHS of formula object}

\item{method}{string: either 'poisson' or 'negbin'}

\item{direction}{string: stepwise direction}

\item{p_threshold}{numeric: threshold for stepwise selection}

\item{k}{integer: limit for k-way interaction terms}

\item{verbose}{logical: whether to print intermediate models}
}
\description{
Helper function for stepwise regression
}
\keyword{internal}
