% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{all_int_cols_to_numeric}
\alias{all_int_cols_to_numeric}
\title{A function used to convert all integer columns to numeric. This is necessary
as the drpop package requires all data columns to be numerics.
This is a temporary workaround until we can implement our own TMLE estimation
function that can handle integer columns.}
\usage{
all_int_cols_to_numeric(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a data frame with all integer columns converted to numeric
}
\description{
A function used to convert all integer columns to numeric. This is necessary
as the drpop package requires all data columns to be numerics.
This is a temporary workaround until we can implement our own TMLE estimation
function that can handle integer columns.
}
\keyword{internal}
