% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control.R
\name{build_individual_records}
\alias{build_individual_records}
\title{Build Individual Simulation Records. This builds all the individual records
for a simulated dataset. Each record produced by this function represents
one individual and contains the capture indicators and one-hot category
indicators. This function is traditionally wrapped by \code{create_data} but
exposed for users who want more granular control over building the simulation
data. Importantly, the capture_probs and category_probs are not validated so
errors may not be as descriptive.}
\usage{
build_individual_records(n_individuals, capture_probs, category_probs)
}
\arguments{
\item{n_individuals}{int: number of participants to simulate}

\item{capture_probs}{numeric vector: capture probabilities of length equal to
the number of capture columns}

\item{category_probs}{numeric vector: category probabilities of length equal
to the number of category columns. Must sum to 1.}
}
\value{
a tibble where each row is an individual record
}
\description{
Build Individual Simulation Records. This builds all the individual records
for a simulated dataset. Each record produced by this function represents
one individual and contains the capture indicators and one-hot category
indicators. This function is traditionally wrapped by \code{create_data} but
exposed for users who want more granular control over building the simulation
data. Importantly, the capture_probs and category_probs are not validated so
errors may not be as descriptive.
}
\keyword{internal}
