% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crc.R
\name{crc}
\alias{crc}
\title{Spatial Capture Re-Capture}
\usage{
crc(data, opts)
}
\arguments{
\item{data}{Dataframe: A dataframe containing a frequency column and binary
columns indicating involvement in the given database}

\item{opts}{Options Object: An object containing the options for the CRC
estimation. This should be an instance of one of the options classes defined
in options.R, such as AICOptions, StepwiseOptions, or EstimatorOptions.}
}
\description{
A method for estimation of 'unknowns' through knowledge about
knowns as described by Barocas, Joshua A et al. "Estimated Prevalence of
Opioid Use Disorder in Massachusetts, 2011-2015: A Capture-Recapture
Analysis." doi:10.2105/AJPH.2018.304673

This implementation supports poisson and negative binomial regression models
with either AIC-based/stepwise formula selection or plugin estimators.
}
