% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crc.R
\name{crc}
\alias{crc}
\title{Spatial Capture Re-Capture}
\usage{
crc(
  data,
  freq_column,
  binary_variables,
  method = "poisson",
  formula_selection = "stepwise",
  formula = NULL,
  opts_stepwise = list(direction = "both", threshold = 0.05, verbose = TRUE)
)
}
\arguments{
\item{data}{Dataframe: A dataframe containing a frequency column and binary
columns indicating involvement in the given database}

\item{freq_column}{Column: A column containing the frequency of observed
combinations}

\item{binary_variables}{List of Columns: List containing columns of binary
variables indicating involvement in the given database}

\item{method}{String: Selection for the spatial capture-recapture method -
either 'poisson' or 'negbin'}

\item{formula_selection}{String: Selection for formula decision - either
'aic' or 'stepwise'}

\item{formula}{Formula: Optional custom formula object for regression}

\item{opts_stepwise}{List: List of \code{direction}: 'forward', 'backward',
or 'both', \code{threshold}: p-value threshold for stepwise selection, and
\code{verbose} if you would like every stepped-through model to be printed}
}
\description{
A method for estimation of 'unknowns' through knowledge about knowns as
described by Barocas, Joshua A et al. "Estimated Prevalence of Opioid Use
Disorder in Massachusetts, 2011-2015: A Capture-Recapture Analysis."
doi:10.2105/AJPH.2018.304673
}
\details{
This implementation supports poisson and negative binomial regression models
with either AIC-based formula selection or stepwise selection.
}
