% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_data}
\alias{create_data}
\title{Data generation tool. This creates a simulated dataset for testing various
Capture-Recapture models. The user is able to specify the number of
individuals, the number of captures, the number of covariate categories
(we currently assume all covariates are one-hot encoded category covariates),
the probabilities of capture for each capture column, and the probabilities
of category assignment for each category column. The returned value
represents the "raw" simulated data where each row is a single sample.
Frequency tables can also be generated by setting
\code{return_frequency_table} to \code{TRUE}, which will aggregate identical
rows and append a frequency column \code{N_ID}.}
\usage{
create_data(
  n_individuals,
  n_captures,
  n_categories = 1,
  p_captures = NULL,
  p_categories = NULL,
  return_frequency_table = FALSE
)
}
\arguments{
\item{n_individuals}{int: number of participants to simulate}

\item{n_captures}{int: number of binary capture columns to simulate}

\item{n_categories}{int: number of one-hot category columns to simulate}

\item{p_captures}{numeric vector: optional capture probabilities of length
\code{n_captures}. Defaults to 0.5 for each capture column.}

\item{p_categories}{numeric vector: optional category probabilities of length
\code{n_categories}. Must sum to 1. Defaults to category_1 = 1.}

\item{return_frequency_table}{logical: when \code{TRUE}, aggregates identical
rows and appends frequency column \code{N_ID}. Default is \code{FALSE}.}
}
\value{
a tibble
}
\description{
Data generation tool. This creates a simulated dataset for testing various
Capture-Recapture models. The user is able to specify the number of
individuals, the number of captures, the number of covariate categories
(we currently assume all covariates are one-hot encoded category covariates),
the probabilities of capture for each capture column, and the probabilities
of category assignment for each category column. The returned value
represents the "raw" simulated data where each row is a single sample.
Frequency tables can also be generated by setting
\code{return_frequency_table} to \code{TRUE}, which will aggregate identical
rows and append a frequency column \code{N_ID}.
}
