% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{is_formula}
\alias{is_formula}
\title{Is Formula}
\usage{
is_formula(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
TRUE if object inherits from class "formula", FALSE otherwise
}
\description{
Generic test of an object being interpretable as a formula
}
